logical/requirements.txt
Pillow>=10.0.0
PyPDF2>=3.0.0
